feat(server): add standard error handler responses
Introduced centralized standard HTTP error responses (400, 404, 405, 500)
to ensure consistent error handling across all server handlers.
Includes helper functions for each response with proper headers and
message bodies.
