+++ internal/server/router.go
@@ -206,43 +206,3 @@ func GetAllowedMethods(methods map[string]HandlerFunc) string {
 	}
 	return strings.Join(allowed, ", ")
 }
-
-// MethodNotAllowedResponse generates a 405 Method Not Allowed response.
-func MethodNotAllowedResponse(allow string) Response {
-	return Response{
-		Version: "HTTP/1.1",
-		Status:  405,
-		Reason:  "Method Not Allowed",
-		Headers: map[string]string{
-			"Content-Type": "text/plain",
-			"Allow":        allow,
-		},
-		Body: []byte("405 Method Not Allowed"),
-	}
-}
-
-// NotFoundResponse generates a 404 Not Found response.
-func NotFoundResponse() Response {
-	return Response{
-		Version: "HTTP/1.1",
-		Status:  404,
-		Reason:  "Not Found",
-		Headers: map[string]string{"Content-Type": "text/plain"},
-		Body:    []byte("404 Not Found"),
-	}
-}
-
-// OptionsResponse generates an automatic 204 No Content OPTIONS response
-// with the Allow header set.
-func OptionsResponse(allow string) Response {
-	utils.Info("Handling automatic OPTIONS response, Allow: %s", allow)
-	return Response{
-		Version: HTTPVersion,
-		Status:  204, // No Content
-		Reason:  "No Content",
-		Headers: map[string]string{
-			"Allow": allow,
-		},
-		Body: []byte(allow),
-	}
-}
